+++ tests/test-runner.js
@@ -23,9 +23,10 @@ import { interactionTests } from "./interactions.test.js";
 import { packageTests } from "./package.test.js";
 import { operationalTests } from "./operational.test.js";
 import { importTests } from "./import.test.js";
+import { tests as productTests } from "./product.test.js";
 
 const resultsEl = document.getElementById("results");
-const tests = [...vizSpecTests, ...vizCompilerTests, ...workspaceTests, ...importTests, ...aiContractTests, ...aiSqlSafetyTests, ...aiContextTests, ...aiProposalTests, ...versionTests, ...dashboardModelTests, ...dashboardFilterTests, ...dashboardExportTests, ...dashboardRunnerTests, ...aiDashboardTests, ...reportModelTests, ...reportRunnerTests, ...reportExportTests, ...aiReportTests, ...mapSpatialTests, ...mapSpecTests, ...mapRendererTests, ...aiMapTests, ...interactionTests, ...packageTests, ...operationalTests];
+const tests = [...vizSpecTests, ...vizCompilerTests, ...workspaceTests, ...importTests, ...productTests, ...aiContractTests, ...aiSqlSafetyTests, ...aiContextTests, ...aiProposalTests, ...versionTests, ...dashboardModelTests, ...dashboardFilterTests, ...dashboardExportTests, ...dashboardRunnerTests, ...aiDashboardTests, ...reportModelTests, ...reportRunnerTests, ...reportExportTests, ...aiReportTests, ...mapSpatialTests, ...mapSpecTests, ...mapRendererTests, ...aiMapTests, ...interactionTests, ...packageTests, ...operationalTests];
 const results = [];
 
 for (const test of tests) {